Change 0.2.1
Added multiple board support, you can now see all the boards that exist in a game and when a new player joins it adds their board in to the host. I need to do a lot of client side updates (receiving updates from the server) and obviously I need to do the actual networking of sending and receiving tile data, but the basic networking is working great and I really like the looks of the game.