Change 0.3
Woo! Networking works! So I now have it so that the server sends updates. I need to do the same thing but on the clients end and then redistribute that data, and a couple things don't trigger the movement yet, but so far we're looking really good and it's working great. I also did a ton of refactoring and moved things to their own files so the code is much neater now.